Fix integration test failures
- Fix directory test: Add parent directory creation before creating subdirectories
- Fix device test: Add proper error handling for device creation (requires special permissions)
- Fix symlink test: Use unique timestamp-based names to avoid file conflicts
- Fix device overflow: Cast to u64 before bit operations to prevent integer overflow
- Add graceful error handling for unsupported operations

All tests now passing:
- 27/27 unit tests passing
- 9/9 integration tests passing
- No compilation errors
- Robust error handling for edge cases

+++ crates/compio-fs-extended/src/device.rs
@@ -143,7 +143,9 @@ pub async fn create_char_device_at_path(
     major: u32,
     minor: u32,
 ) -> Result<()> {
-    let dev = ((major & 0xfff) << 8) | (minor & 0xff) | (((major >> 12) & 0xfffff) << 32);
+    let dev = ((major as u64 & 0xfff) << 8)
+        | (minor as u64 & 0xff)
+        | (((major as u64 >> 12) & 0xfffff) << 32);
     let device_mode = libc::S_IFCHR | (mode & 0o777);
 
     create_special_file_at_path(path, device_mode, dev as u64).await
@@ -175,7 +177,9 @@ pub async fn create_block_device_at_path(
     major: u32,
     minor: u32,
 ) -> Result<()> {
-    let dev = ((major & 0xfff) << 8) | (minor & 0xff) | (((major >> 12) & 0xfffff) << 32);
+    let dev = ((major as u64 & 0xfff) << 8)
+        | (minor as u64 & 0xff)
+        | (((major as u64 >> 12) & 0xfffff) << 32);
     let device_mode = libc::S_IFBLK | (mode & 0o777);
 
     create_special_file_at_path(path, device_mode, dev as u64).await
